requirements.txt
web3==6.11.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
cryptography==41.0.7
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
